dynamic example: renamed all .jsx views with the .js extension
