+++ examples/dynamic/package.json
@@ -6,7 +6,7 @@
   "private": true,
   "scripts": {
     "preinstall": "npm install ../../",
-    "start": "browserify -t reactify views/main.jsx -o public/main.js; node app.js"
+    "start": "browserify -t reactify views/main.js -o public/main.js; node app.js"
   },
   "dependencies": {
     "browserify": "^8.0.3",
