+++ examples/dynamic/views/Content.js
