+++ examples/dynamic/views/main.js
@@ -1,5 +1,5 @@
 var React = require('react');
-var Content = require('./Content.jsx');
+var Content = require('./Content');
 
 window.main = function (data, containerId) {
   var container = document.getElementById(containerId || 'content');
